requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic
pydantic-settings

python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2

python-multipart

alembic
psycopg2-binary
